requirements.txt
numpy
matplotlib
torch
transformers
datasets
scikit-learn
fastapi
uvicorn
pydantic
google-genai
pytest
python-dotenv
